requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
numpy==1.23.1
opencv-python==4.6.0.66
pytz==2022.1
sqlparse==0.4.2
